feat(core): cover the simple cases too
