+++ harper-core/src/patterns/sequence_pattern.rs
@@ -160,9 +160,9 @@ impl SequencePattern {
         self
     }
 
-    pub fn then_one_or_more(mut self, pat: Box<dyn Pattern>) -> Self {
+    pub fn then_one_or_more(mut self, pat: impl Pattern + 'static) -> Self {
         self.token_patterns
-            .push(Box::new(RepeatingPattern::new(pat, 0)));
+            .push(Box::new(RepeatingPattern::new(Box::new(pat), 0)));
         self
     }
